requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
groq
python-dotenv
jinja2
aiofiles
